Chore: extract user Id util rename [#229]
