+++ src/utils/index.ts
@@ -1,4 +1,4 @@
-import extractUserId from './extractNickname';
+import extractUserId from './extractUserId';
 import { handleScrollDown, handleScrollUp } from './handleScroll';
 
 export { handleScrollDown, handleScrollUp, extractUserId };
